requirements.txt
# ======================================================================
# Core dependencies for InstructPix2Pix training
# ======================================================================
# ⚠️ 重要：请按照以下步骤安装环境
# 1. 创建 conda 环境：conda create -n dl-final python=3.10 -y
# 2. 激活环境：conda activate dl-final
# 3. 安装依赖：pip install -r requirements.txt

# ======================================================================
# PyTorch 及其生态系统
# ======================================================================
torch>=2.0.0
torchvision>=0.15.0

# ======================================================================
# HuggingFace 生态系统
# ======================================================================
accelerate>=0.20.0          # 分布式训练支持
datasets>=2.0.0             # 数据集处理（用于 make_dataset.py）
diffusers>=0.35.0,<0.36.0   # 扩散模型库（InstructPix2Pix）
transformers>=4.25.0        # 模型库（CLIP 文本编码器）
huggingface_hub>=0.16.0     # HuggingFace Hub 集成（模型下载）

# ======================================================================
# 图像处理
# ======================================================================
Pillow>=9.0.0               # 图像处理（用于 create_mini_data.py, make_dataset.py）
numpy>=1.21.0               # 数值计算

# ======================================================================
# 实用工具
# ======================================================================
tqdm>=4.64.0                # 进度条显示
packaging>=21.0             # 版本检查工具
requests>=2.28.0            # HTTP 请求

# ======================================================================
# 可选依赖（推荐安装）
# ======================================================================
wandb>=0.13.0               # 实验跟踪和可视化（可选但推荐）

# ======================================================================
# 兼容性修复
# ======================================================================
# Fix protobuf compatibility (required for torch/onnx compatibility)
protobuf<=3.20.3
# Fix pyarrow compatibility for cudf
pyarrow>=14.0.1,<15.0.0

# ======================================================================
# 可选：内存优化（根据需要单独安装）
# ======================================================================
# xformers>=0.0.17          # 内存高效的注意力机制（取消注释以启用）
# bitsandbytes>=0.37.0     # 8-bit Adam 优化器（取消注释以启用）

# ======================================================================
# 使用镜像站点（加速下载）
# ======================================================================
# 如果无法访问 huggingface.co，可使用镜像：
#   export HF_ENDPOINT=https://hf-mirror.com

